+++ test-app/src/global.d.ts
@@ -1,16 +1,23 @@
 /// <reference types="@sveltejs/kit" />
 
-declare module '*.svg?src' {
-  const content: string
+declare module '*.svg?component' {
+  import { SvelteComponent } from 'svelte'
+  const src: SvelteComponent
   export default content
 }
 
-declare module '*.svg?component' {
+declare module '*.svg?src' {
   const content: string
   export default content
 }
 
 declare module '*.svg?url' {
   const content: string
+  export default content
+}
+
+declare module '*.svg' {
+  import { SvelteComponent } from 'svelte'
+  const src: SvelteComponent
   export default content
 }